feat: prevent loading state while waiting for the post
